Basic functionality is down
Gonna be real with yall, i have no idea exactly what changed here, but it works pretty well and what i have so far is merely for testing purposes.

I believe its just a matter of setting up the game now that that is done, maybe should add the text animation because its very rough at the moment, but its not necessary.